Moved DatabaseloggerPrivat into own class header and source file.
Databaselogger and DatabaseloggerPrivate are both quite big classes.
We split them here for better readability.
