+++ vl_databaselogger.cpp
@@ -16,441 +16,13 @@
 #include <vcmp_errordata.h>
 #include <veinmodulerpc.h>
 #include <QJsonDocument>
+#include <vl_databaseloggerprivate.h>
 
 Q_LOGGING_CATEGORY(VEIN_LOGGER, VEIN_DEBUGNAME_LOGGER)
 
 namespace VeinLogger
 {
-class DataLoggerPrivate: public QObject
-{
-    explicit DataLoggerPrivate(DatabaseLogger *t_qPtr) : m_qPtr(t_qPtr)
-    {
-        m_batchedExecutionTimer.setInterval(5000);
-        m_batchedExecutionTimer.setSingleShot(false);
-    }
-    ~DataLoggerPrivate()
-    {
-        m_batchedExecutionTimer.stop();
-        if(m_database != nullptr)
-        {
-            m_database->deleteLater(); ///@todo: check if the delete works across threads
-            m_database = nullptr;
-        }
-        m_asyncDatabaseThread.quit();
-        m_asyncDatabaseThread.wait();
-    }
 
-    void initOnce() {
-        Q_ASSERT(m_initDone == false);
-        if(m_initDone == false) {
-            VeinComponent::EntityData *systemData = new VeinComponent::EntityData();
-            systemData->setCommand(VeinComponent::EntityData::Command::ECMD_ADD);
-            systemData->setEntityId(m_entityId);
-
-            VeinEvent::CommandEvent *systemEvent = new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, systemData);
-
-            emit m_qPtr->sigSendEvent(systemEvent);
-
-            VeinComponent::ComponentData *initialData = nullptr;
-
-            QHash<QString, QVariant> componentData;
-            componentData.insert(s_entityNameComponentName, m_entityName);
-            componentData.insert(s_loggingEnabledComponentName, QVariant(false));
-            componentData.insert(s_loggingStatusTextComponentName, QVariant(QString("Logging inactive")));
-            ///@todo load from persistent settings file?
-            componentData.insert(s_databaseReadyComponentName, QVariant(false));
-            componentData.insert(s_databaseFileComponentName, QVariant(QString()));
-            componentData.insert(s_databaseErrorFileComponentName, QVariant(QString()));
-            componentData.insert(s_databaseFileMimeTypeComponentName, QVariant(QString()));
-            componentData.insert(s_databaseFileSizeComponentName, QVariant(QString()));
-            componentData.insert(s_filesystemInfoComponentName, QVariantMap());
-            componentData.insert(s_scheduledLoggingEnabledComponentName, QVariant(false));
-            componentData.insert(s_scheduledLoggingDurationComponentName, QVariant());
-            componentData.insert(s_scheduledLoggingCountdownComponentName, QVariant(0.0));
-            componentData.insert(s_existingSessionsComponentName, QStringList());
-            componentData.insert(s_customerDataComponentName, QString());
-
-            // TODO: Add more from modulemanager
-            componentData.insert(s_sessionNameComponentName, QString());
-            componentData.insert(s_guiContextComponentName, QString());
-            componentData.insert(s_transactionNameComponentName, QString());
-            componentData.insert(s_currentContentSetsComponentName, QStringList());
-            componentData.insert(s_availableContentSetsComponentName, QStringList());
-
-            for(const QString &componentName : componentData.keys()) {
-                initialData = new VeinComponent::ComponentData();
-                initialData->setEntityId(m_entityId);
-                initialData->setCommand(VeinComponent::ComponentData::Command::CCMD_ADD);
-                initialData->setComponentName(componentName);
-                initialData->setNewValue(componentData.value(componentName));
-                initialData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-                initialData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-
-                systemEvent = new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, initialData);
-                emit m_qPtr->sigSendEvent(systemEvent);
-            }
-
-            QMap<QString,QString> tmpParamMap;
-            VfCpp::cVeinModuleRpc::Ptr tmpval= VfCpp::cVeinModuleRpc::Ptr(new VfCpp::cVeinModuleRpc(m_entityId,m_qPtr,m_qPtr,"RPC_readTransaction",VfCpp::cVeinModuleRpc::Param({{"p_session", "QString"},{"p_transaction", "QString"}})), &QObject::deleteLater);
-            m_rpcList[tmpval->rpcName()]=tmpval;
-            tmpval= VfCpp::cVeinModuleRpc::Ptr(new VfCpp::cVeinModuleRpc(m_entityId,m_qPtr,m_qPtr,"RPC_readSessionComponent",VfCpp::cVeinModuleRpc::Param({{"p_session", "QString"},{"p_entity", "QString"},{"p_component", "QString"}})), &QObject::deleteLater);
-            m_rpcList[tmpval->rpcName()]=tmpval;
-            tmpval= VfCpp::cVeinModuleRpc::Ptr(new VfCpp::cVeinModuleRpc(m_entityId,m_qPtr,m_qPtr,"RPC_deleteSession",VfCpp::cVeinModuleRpc::Param({{"p_session", "QString"}})), &QObject::deleteLater);
-            m_rpcList[tmpval->rpcName()]=tmpval;
-
-
-            initStateMachine();
-
-            m_initDone = true;
-        }
-    }
-
-    void setStatusText(const QString &t_status)
-    {
-        if(m_loggerStatusText != t_status) {
-            m_loggerStatusText = t_status;
-
-            VeinComponent::ComponentData *schedulingEnabledData = new VeinComponent::ComponentData();
-            schedulingEnabledData->setEntityId(m_entityId);
-            schedulingEnabledData->setCommand(VeinComponent::ComponentData::Command::CCMD_SET);
-            schedulingEnabledData->setComponentName(DataLoggerPrivate::s_loggingStatusTextComponentName);
-            schedulingEnabledData->setNewValue(t_status);
-            schedulingEnabledData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-            schedulingEnabledData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-
-            emit m_qPtr->sigSendEvent(new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, schedulingEnabledData));
-        }
-    }
-
-    void initStateMachine()
-    {
-        m_parallelWrapperState->setChildMode(QStateMachine::ParallelStates);
-        m_stateMachine.setInitialState(m_parallelWrapperState);
-        m_databaseContainerState->setInitialState(m_databaseUninitializedState);
-        m_loggingContainerState->setInitialState(m_loggingDisabledState);
-        m_logSchedulerContainerState->setInitialState(m_logSchedulerDisabledState);
-
-        //uninitialized -> ready
-        m_databaseUninitializedState->addTransition(m_qPtr, &DatabaseLogger::sigDatabaseReady, m_databaseReadyState);
-        //ready -> uninitialized
-        m_databaseReadyState->addTransition(m_qPtr, &DatabaseLogger::sigDatabaseUnloaded, m_databaseUninitializedState);
-
-        //enabled -> disabled
-        m_loggingEnabledState->addTransition(m_qPtr, &DatabaseLogger::sigLoggingStopped, m_loggingDisabledState);
-        //disabled -> enabled
-        m_loggingDisabledState->addTransition(m_qPtr, &DatabaseLogger::sigLoggingStarted, m_loggingEnabledState);
-
-        //enabled -> disbled
-        m_logSchedulerEnabledState->addTransition(m_qPtr, &DatabaseLogger::sigLogSchedulerDeactivated, m_logSchedulerDisabledState);
-        //disabled -> enabled
-        m_logSchedulerDisabledState->addTransition(m_qPtr, &DatabaseLogger::sigLogSchedulerActivated, m_logSchedulerEnabledState);
-
-        QObject::connect(m_databaseUninitializedState, &QState::entered, [&]() {
-            VeinComponent::ComponentData *databaseUninitializedCData = new VeinComponent::ComponentData();
-            databaseUninitializedCData->setEntityId(m_entityId);
-            databaseUninitializedCData->setCommand(VeinComponent::ComponentData::Command::CCMD_SET);
-            databaseUninitializedCData->setComponentName(DataLoggerPrivate::s_databaseReadyComponentName);
-            databaseUninitializedCData->setNewValue(false);
-            databaseUninitializedCData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-            databaseUninitializedCData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-
-            emit m_qPtr->sigSendEvent(new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, databaseUninitializedCData));
-            m_qPtr->setLoggingEnabled(false);
-            if(!m_noUninitMessage) {
-                setStatusText("No database selected");
-            }
-        });
-        QObject::connect(m_databaseReadyState, &QState::entered, [&](){
-            // Now we have an open and valid database: Notify ready and some
-            // bits we were not sure to have at openDatabase
-            QHash <QString, QVariant> fileInfoData;
-            QFileInfo fileInfo(m_databaseFilePath);
-            QMimeDatabase mimeDB;
-            fileInfoData.insert(DataLoggerPrivate::s_databaseFileMimeTypeComponentName, mimeDB.mimeTypeForFile(fileInfo, QMimeDatabase::MatchContent).name());
-            fileInfoData.insert(DataLoggerPrivate::s_databaseFileSizeComponentName, fileInfo.size());
-            fileInfoData.insert(DataLoggerPrivate::s_databaseReadyComponentName, true);
-            for(const QString &componentName : fileInfoData.keys())  {
-                VeinComponent::ComponentData *storageCData = new VeinComponent::ComponentData();
-                storageCData->setEntityId(m_entityId);
-                storageCData->setCommand(VeinComponent::ComponentData::Command::CCMD_SET);
-                storageCData->setComponentName(componentName);
-                storageCData->setNewValue(fileInfoData.value(componentName));
-                storageCData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-                storageCData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-                emit m_qPtr->sigSendEvent(new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, storageCData));
-            }
-
-            m_qPtr->setLoggingEnabled(false);
-            setStatusText("Database loaded");
-
-            QStorageInfo storageInfo(fileInfo.absolutePath());
-            // * To avoid fire storm on logging we watch file's dir
-            // * For removable devices: mount-point's parent dir
-            QStringList watchedPaths;
-            watchedPaths.append(fileInfo.absolutePath());
-            if(!storageInfo.isRoot()) {
-                QDir tmpDir(storageInfo.rootPath());
-                tmpDir.cdUp();
-                watchedPaths.append(tmpDir.path());
-            }
-            qInfo("Database logger watching path(s): %s", qPrintable(watchedPaths.join(QStringLiteral(" + "))));
-            QStringList unWatchedPaths = m_deleteWatcher.addPaths(watchedPaths);
-            if(m_deleteWatcher.directories().count()) {
-                QObject::connect(&m_deleteWatcher, &QFileSystemWatcher::directoryChanged, m_qPtr, &DatabaseLogger::checkDatabaseStillValid);
-            }
-            if(unWatchedPaths.count()) {
-                qWarning("Unwatched paths: %s", qPrintable(unWatchedPaths.join(QStringLiteral(" + "))));
-            }
-        });
-        QObject::connect(m_loggingEnabledState, &QState::entered, [&](){
-            setStatusText("Logging data");
-        });
-        QObject::connect(m_loggingDisabledState, &QState::entered, [&](){
-            if(!m_noUninitMessage) {
-                // yes we are in logging disabled state - but the message 'Logging disabled' is a
-                // bit misleading: sounds as something is wrong and blocking further logging
-                //setStatusText("Logging disabled");
-                setStatusText("Database loaded");
-            }
-            m_batchedExecutionTimer.stop();
-            updateDBFileSizeInfo();
-        });
-        QObject::connect(m_logSchedulerEnabledState, &QState::entered, [&](){
-            VeinComponent::ComponentData *schedulingEnabledCData = new VeinComponent::ComponentData();
-            schedulingEnabledCData->setEntityId(m_entityId);
-            schedulingEnabledCData->setCommand(VeinComponent::ComponentData::Command::CCMD_SET);
-            schedulingEnabledCData->setComponentName(DataLoggerPrivate::s_scheduledLoggingEnabledComponentName);
-            schedulingEnabledCData->setNewValue(true);
-            schedulingEnabledCData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-            schedulingEnabledCData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-
-            emit m_qPtr->sigSendEvent(new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, schedulingEnabledCData));
-        });
-        QObject::connect(m_logSchedulerDisabledState, &QState::entered, [&](){
-            VeinComponent::ComponentData *schedulingDisabledCData = new VeinComponent::ComponentData();
-            schedulingDisabledCData->setEntityId(m_entityId);
-            schedulingDisabledCData->setCommand(VeinComponent::ComponentData::Command::CCMD_SET);
-            schedulingDisabledCData->setComponentName(DataLoggerPrivate::s_scheduledLoggingEnabledComponentName);
-            schedulingDisabledCData->setNewValue(false);
-            schedulingDisabledCData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-            schedulingDisabledCData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-
-            emit m_qPtr->sigSendEvent(new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, schedulingDisabledCData));
-        });
-
-        m_stateMachine.start();
-    }
-
-    void updateDBStorageInfo()
-    {
-        const auto storages = QStorageInfo::mountedVolumes();
-        QVariantMap storageInfoMap;
-        for(const auto storDevice : storages)  {
-            if(storDevice.fileSystemType().contains("tmpfs") == false) {
-                const double availGB = storDevice.bytesFree()/1.0e9;
-                const double totalGB = storDevice.bytesTotal()/1.0e9;
-
-                QVariantMap storageData;
-                storageData.insert(DataLoggerPrivate::s_filesystemFreePropertyName, availGB);
-                storageData.insert(DataLoggerPrivate::s_filesystemTotalPropertyName, totalGB);
-
-                storageInfoMap.insert(storDevice.rootPath(), storageData);
-            }
-        }
-
-        VeinComponent::ComponentData *storageCData= new VeinComponent::ComponentData();
-        storageCData->setEntityId(m_entityId);
-        storageCData->setCommand(VeinComponent::ComponentData::Command::CCMD_SET);
-        storageCData->setComponentName(s_filesystemInfoComponentName);
-        storageCData->setNewValue(storageInfoMap);
-        storageCData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-        storageCData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-
-        emit m_qPtr->sigSendEvent(new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, storageCData));
-    }
-
-    bool checkDBFilePath(const QString &t_dbFilePath)
-    {
-        bool retVal = false;
-        QFileInfo fInfo(t_dbFilePath);
-
-        if(!fInfo.isRelative()) {
-            // try to create path
-            if(!fInfo.absoluteDir().exists()) {
-                QDir dir;
-                dir.mkpath(fInfo.absoluteDir().path());
-            }
-
-            if(fInfo.absoluteDir().exists()) {
-                if(fInfo.isFile() || fInfo.exists() == false) {
-                    retVal = true;
-                }
-                else {
-                    emit m_qPtr->sigDatabaseError(QString("Path is not a valid file location: %1").arg(t_dbFilePath));
-                }
-            }
-            else {
-                emit m_qPtr->sigDatabaseError(QString("Parent directory for path does not exist: %1").arg(t_dbFilePath));
-            }
-        }
-        else {
-            emit m_qPtr->sigDatabaseError(QString("Relative paths are not accepted: %1").arg(t_dbFilePath));
-        }
-
-        return retVal;
-    }
-
-    void updateDBFileSizeInfo()
-    {
-        QFileInfo fInfo(m_databaseFilePath);
-        if(fInfo.exists()) {
-            VeinComponent::ComponentData *storageCData = new VeinComponent::ComponentData();
-            storageCData->setEntityId(m_entityId);
-            storageCData->setCommand(VeinComponent::ComponentData::Command::CCMD_SET);
-            storageCData->setComponentName(DataLoggerPrivate::s_databaseFileSizeComponentName);
-            storageCData->setNewValue(QVariant(fInfo.size()));
-            storageCData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-            storageCData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-
-            emit m_qPtr->sigSendEvent(new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, storageCData));
-        }
-    }
-
-    void updateSchedulerCountdown()
-    {
-        if(m_schedulingTimer.isActive()) {
-            VeinComponent::ComponentData *schedulerCountdownCData = new VeinComponent::ComponentData();
-            schedulerCountdownCData->setEntityId(m_entityId);
-            schedulerCountdownCData->setCommand(VeinComponent::ComponentData::Command::CCMD_SET);
-            schedulerCountdownCData->setComponentName(DataLoggerPrivate::s_scheduledLoggingCountdownComponentName);
-            schedulerCountdownCData->setNewValue(QVariant(m_schedulingTimer.remainingTime()));
-            schedulerCountdownCData->setEventOrigin(VeinEvent::EventData::EventOrigin::EO_LOCAL);
-            schedulerCountdownCData->setEventTarget(VeinEvent::EventData::EventTarget::ET_ALL);
-
-            emit m_qPtr->sigSendEvent(new VeinEvent::CommandEvent(VeinEvent::CommandEvent::EventSubtype::NOTIFICATION, schedulerCountdownCData));
-        }
-    }
-
-    /**
-     * @brief The logging is implemented via interpreted scripts that state which values to log
-     * @see vl_qmllogger.cpp
-     */
-    QVector<QmlLogger *> m_loggerScripts;
-
-    /**
-     * @brief The actual database choice is an implementation detail of the DatabaseLogger
-     */
-    AbstractLoggerDB *m_database=nullptr;
-    DBFactory m_databaseFactory;
-    QString m_databaseFilePath;
-    DataSource *m_dataSource=nullptr;
-
-    /**
-     * @brief Qt doesn't support non blocking database access
-     */
-    QThread m_asyncDatabaseThread;
-    /**
-     * @b Logging in batches is much more efficient for SQLITE (and for spinning disk storages in general)
-     * @note The batch timer is independent from the recording timeframe as it only pushes already logged values to the database
-     */
-    QTimer m_batchedExecutionTimer;
-    /**
-     * @brief logging duration in ms
-     */
-    int m_scheduledLoggingDuration;
-
-    QFileSystemWatcher m_deleteWatcher;
-    bool m_noUninitMessage = false;
-
-    QTimer m_schedulingTimer;
-    QTimer m_countdownUpdateTimer;
-    bool m_initDone=false;
-    QString m_loggerStatusText="Logging inactive";
-
-    QMap<QString,VfCpp::cVeinModuleRpc::Ptr> m_rpcList;
-    /**
-     * @brief m_sessionName
-     * stores the current session Name.
-     *
-     * We need this, when deleting a session, to compare values.
-     * That's a quite ugly solution. However using vf-cpp in future those things
-     * will work out better.
-     */
-    QString m_sessionName;
-
-    int m_entityId;
-    //entity name
-    QLatin1String m_entityName;
-    //component names
-    static constexpr QLatin1String s_entityNameComponentName = QLatin1String("EntityName");
-    static constexpr QLatin1String s_loggingStatusTextComponentName = QLatin1String("LoggingStatus");
-    static constexpr QLatin1String s_loggingEnabledComponentName = QLatin1String("LoggingEnabled");
-    static constexpr QLatin1String s_databaseReadyComponentName = QLatin1String("DatabaseReady");
-    static constexpr QLatin1String s_databaseFileComponentName = QLatin1String("DatabaseFile");
-    static constexpr QLatin1String s_databaseErrorFileComponentName = QLatin1String("DatabaseErrorFile");
-    static constexpr QLatin1String s_databaseFileMimeTypeComponentName = QLatin1String("DatabaseFileMimeType");
-    static constexpr QLatin1String s_databaseFileSizeComponentName = QLatin1String("DatabaseFileSize");
-    static constexpr QLatin1String s_filesystemInfoComponentName = QLatin1String("FilesystemInfo");
-    static constexpr QLatin1String s_filesystemFreePropertyName = QLatin1String("FilesystemFree");
-    static constexpr QLatin1String s_filesystemTotalPropertyName = QLatin1String("FilesystemTotal");
-    static constexpr QLatin1String s_scheduledLoggingEnabledComponentName = QLatin1String("ScheduledLoggingEnabled");
-    static constexpr QLatin1String s_scheduledLoggingDurationComponentName = QLatin1String("ScheduledLoggingDuration");
-    static constexpr QLatin1String s_scheduledLoggingCountdownComponentName = QLatin1String("ScheduledLoggingCountdown");
-    static constexpr QLatin1String s_existingSessionsComponentName = QLatin1String("ExistingSessions");
-    static constexpr QLatin1String s_customerDataComponentName = QLatin1String("CustomerData");
-
-    // TODO: Add more from modulemanager
-    static constexpr QLatin1String s_sessionNameComponentName = QLatin1String("sessionName");
-    static constexpr QLatin1String s_guiContextComponentName = QLatin1String("guiContext");
-    static constexpr QLatin1String s_transactionNameComponentName = QLatin1String("transactionName");
-    static constexpr QLatin1String s_currentContentSetsComponentName = QLatin1String("currentContentSets");
-    static constexpr QLatin1String s_availableContentSetsComponentName = QLatin1String("availableContentSets");
-
-    QStateMachine m_stateMachine;
-    //QStatemachine does not support ParallelState
-    //Therefore we add one state where the actual statemachine starts
-    QState *m_parallelWrapperState = new QState(&m_stateMachine);
-
-    QState *m_databaseContainerState = new QState(m_parallelWrapperState);
-    QState *m_databaseUninitializedState = new QState(m_databaseContainerState);
-    QState *m_databaseReadyState = new QState(m_databaseContainerState);
-
-    QState *m_loggingContainerState = new QState(m_parallelWrapperState);
-    QState *m_loggingEnabledState = new QState(m_loggingContainerState);
-    QState *m_loggingDisabledState = new QState(m_loggingContainerState);
-
-    QState *m_logSchedulerContainerState = new QState(m_parallelWrapperState);
-    QState *m_logSchedulerEnabledState = new QState(m_logSchedulerContainerState);
-    QState *m_logSchedulerDisabledState = new QState(m_logSchedulerContainerState);
-
-    AbstractLoggerDB::STORAGE_MODE m_storageMode;
-
-    DatabaseLogger *m_qPtr=nullptr;
-    friend class DatabaseLogger;
-};
-//constexpr definition, see: https://stackoverflow.com/questions/8016780/undefined-reference-to-static-constexpr-char
-constexpr QLatin1String DataLoggerPrivate::s_entityNameComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_loggingStatusTextComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_loggingEnabledComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_databaseReadyComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_databaseFileComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_databaseErrorFileComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_databaseFileMimeTypeComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_databaseFileSizeComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_filesystemInfoComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_filesystemFreePropertyName;
-constexpr QLatin1String DataLoggerPrivate::s_filesystemTotalPropertyName;
-constexpr QLatin1String DataLoggerPrivate::s_scheduledLoggingEnabledComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_scheduledLoggingDurationComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_scheduledLoggingCountdownComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_existingSessionsComponentName;
-// TODO: Add more from modulemanager
-constexpr QLatin1String DataLoggerPrivate::s_customerDataComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_sessionNameComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_guiContextComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_transactionNameComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_currentContentSetsComponentName;
-constexpr QLatin1String DataLoggerPrivate::s_availableContentSetsComponentName;
 
 DatabaseLogger::DatabaseLogger(DataSource *t_dataSource, DBFactory t_factoryFunction, QObject *t_parent, AbstractLoggerDB::STORAGE_MODE t_storageMode) :
     VeinEvent::EventSystem(t_parent),
